bug: Requires .so file to run
Tried to fix using cython, but I realized I need to figure out how
to link the file myself (which I have no experience doing! the
target machine doesn't even have a working gcc LOL).

Committing my changes in case I want to work from this point in the
future, but for now I'm going to scale back on the project and try
to use the cron job to hide.